pyproject.toml
[project]
name = "test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-adk>=0.2.0",
    "jupyterlab>=4.4.0",
    "numpy>=2.2.4",
    "pandas>=2.2.3",
]
